removing TODO's + renaming files for consistency
