+++ src/test/overlay.test.tsx
@@ -4,7 +4,7 @@ import { act, renderHook } from "@testing-library/react-hooks";
 import Overlay from "../components/Editor/Overlay";
 import { useActionList, useOverlay } from "../util/customHooks/overlayfunc";
 import { ActionData } from "../util/types";
-import { useActionBar } from "../util/customHooks/ActionBarFunc";
+import { useActionBar } from "../util/customHooks/actionBarFunc";
 import {
   globalContextTypes,
   GlobalContext,
